requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents
azure-identity
python-dotenv
